requirements.txt
argcomplete==3.1.4
asgiref==3.4.1
attrs==23.2.0
autopep8==1.6.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
build==1.2.2.post1
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
click==8.1.7
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==43.0.3
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
distro-info==1.7+build1
Django==3.2.13
django-ckeditor==6.3.0
django-ckeditor-5==0.2.13
django-cors-headers==3.11.0
django-countries==7.6.1
django-crispy-forms==1.13.0
django-import-export==2.5.0
django-jazzmin==2.4.2
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-taggit==1.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
duplicity==2.1.4
et-xmlfile==1.1.0
fasteners==0.18
gunicorn==23.0.0
gyp==0.1
httplib2==0.20.4
idna==3.6
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
mysql-connector-python==9.0.0
mysqlclient==2.2.4
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.5
packaging==24.1
paramiko==2.12.0
pexpect==4.9.0
Pillow==9.0.1
pip-tools==7.4.1
pipx==1.4.3
platformdirs==4.2.0
psutil==5.9.8
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycairo==1.25.1
pycodestyle==2.8.0
pycparser==2.22
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.3.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.1.1
pyproject_hooks==1.2.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==0.19.2
pytz==2021.3
pyxdg==0.28
PyYAML==6.0.2
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.4.2
systemd-python==235
tablib==3.6.1
toml==0.10.2
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2021.5
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
wadllib==1.3.6
watchdog==5.0.3
whitenoise==6.7.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
